backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
reportlab==4.0.7
python-dateutil==2.8.2
openai==1.54.0
python-dotenv==1.0.0
PyPDF2==3.0.1
requests==2.31.0
